Moving Towards Modules, Starting With Messages
The code was starting to get unwieldy, so I'm going to begin moving it
all into proper modules.  The first to get this treatment was the
message-passing code, which got consolidated into the Transceiver
module.  It defines a simpler interface for the background process to
set up a server and the popups, option pages and content scripts to
send/receive messages to/from it.  It also lets me add local
interceptors (to update the browser action icons, for instance)
without hacking them into all the callsites that send the messages.