Add Storage Method Setting to Options Page
This turned out to be significantly more difficult than expected.
I've added a control on the options page to let the user switch back
and forth between using local and sync storage for their passwords.
When they choose to switch, they're also given the option to transfer
their passwords to the new storage method if desired.